Add comments and implement a more robust approach to detect parsing errors.
Instead of checking for the 3 conditions (severity, message and ruleId), just check the ruleId.

It is not completely possible to be safe from any change in ESLint -for example, if they decide
to assign a ruleId for parsing errors instead of using null this check would break-
but checking only the ruleId seems the least fragile of the conditions and the string prepended
to the message could change from "Parsing error:" to "Parsing error - ", "Error during parsing", etc.